requirements.txt
appnope==0.1.4
async_lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
cellpose==4.1.1
certifi==2026.2.25
cffi==2.0.0
charset_normalizer==3.4.7
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.20
defusedxml==0.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastjsonschema==2.21.2
fastremap==1.18.0
filelock==3.28.0
fill_voids==2.1.1
fontTools==4.62.1
fqdn==1.5.1
fsspec==2026.3.0
graphtools==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.3
ipykernel==7.2.0
IPython==8.39.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonschema_specifications==2025.9.1
jupyter_client==8.8.0
jupyter_console==6.6.3
jupyter_core==5.9.1
jupyter_events==0.12.0
jupyter_lsp==2.3.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.5.0
lazy_loader==0.5
llvmlite==0.47.0
markupsafe==3.0.3
matplotlib_inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
networkx==3.4.2
notebook_shim==0.2.4
numba==0.65.0
opencv-python==4.13.0.92
openpyxl==3.1.5
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
Pillow==12.2.0
pip==24.3.1
platformdirs==4.9.6
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycparser==3.0
pygsp==0.6.1
pynndescent==0.6.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rfc3987_syntax==1.1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
segment_anything==1.0
send2trash==2.1.0
setuptools==44.1.1
soupsieve==2.8.3
stack_data==0.6.3
sympy==1.14.0
tasklogger==1.2.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.4.1
torchaudio==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
umap-learn==0.5.12
uri_template==1.3.0
urllib3==2.6.3
webcolors==25.10.0
webencodings==0.5.1
wrapt==2.1.2
